refactor(arel): inline the Array to_s helper into the to-sql visitor
